Backend_Django/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-extensions==2.2.1
djangorestframework==3.10.2
idna==2.8
IMDbPY==6.8
joblib==0.13.2
lxml==4.4.1
numpy==1.17.0
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.9
sqlparse==0.3.0
urllib3==1.25.3
